admin api for storageinfo
