+++ api/server/admin/storageinfo.go
@@ -1,11 +1,13 @@
-package user
+package admin
 
 import (
 	"context"
 
 	"github.com/ipfs/go-cid"
+	adminPb "github.com/textileio/powergate/api/gen/powergate/admin/v1"
 	userPb "github.com/textileio/powergate/api/gen/powergate/user/v1"
 	su "github.com/textileio/powergate/api/server/util"
+	"github.com/textileio/powergate/ffs"
 	"github.com/textileio/powergate/ffs/api"
 	"github.com/textileio/powergate/util"
 	"google.golang.org/grpc/codes"
@@ -14,16 +16,13 @@ import (
 
 // StorageInfo returns the information about a stored Cid. If no information is available,
 // since the Cid was never stored, it returns an error with codes.NotFound.
-func (s *Service) StorageInfo(ctx context.Context, req *userPb.StorageInfoRequest) (*userPb.StorageInfoResponse, error) {
-	i, err := s.getInstanceByToken(ctx)
-	if err != nil {
-		return nil, status.Errorf(codes.PermissionDenied, "getting instance: %v", err)
-	}
+func (a *Service) StorageInfo(ctx context.Context, req *adminPb.StorageInfoRequest) (*adminPb.StorageInfoResponse, error) {
+	iid := ffs.APIID(req.UserId)
 	cid, err := util.CidFromString(req.Cid)
 	if err != nil {
 		return nil, status.Errorf(codes.InvalidArgument, "parsing cid: %v", err)
 	}
-	info, err := i.StorageInfo(cid)
+	info, err := a.s.GetStorageInfo(iid, cid)
 	if err == api.ErrNotFound {
 		return nil, status.Errorf(codes.NotFound, "getting storage info: %v", err)
 	}
@@ -31,14 +30,14 @@ func (s *Service) StorageInfo(ctx context.Context, req *userPb.StorageInfoReques
 		return nil, status.Errorf(codes.Internal, "getting storage info: %v", err)
 	}
 	pbInfo := su.ToRPCStorageInfo(info)
-	return &userPb.StorageInfoResponse{StorageInfo: pbInfo}, nil
+	return &adminPb.StorageInfoResponse{StorageInfo: pbInfo}, nil
 }
 
-// ListStorageInfo returns a list of information about all stored cids, filtered by cids if provided.
-func (s *Service) ListStorageInfo(ctx context.Context, req *userPb.ListStorageInfoRequest) (*userPb.ListStorageInfoResponse, error) {
-	i, err := s.getInstanceByToken(ctx)
-	if err != nil {
-		return nil, status.Errorf(codes.PermissionDenied, "getting instance: %v", err)
+// ListStorageInfo returns a list of information about all stored cids, filtered by user ids and cids if provided.
+func (a *Service) ListStorageInfo(ctx context.Context, req *adminPb.ListStorageInfoRequest) (*adminPb.ListStorageInfoResponse, error) {
+	iids := make([]ffs.APIID, len(req.UserIds))
+	for i, id := range req.UserIds {
+		iids[i] = ffs.APIID(id)
 	}
 	cids := make([]cid.Cid, len(req.Cids))
 	for i, s := range req.Cids {
@@ -48,7 +47,7 @@ func (s *Service) ListStorageInfo(ctx context.Context, req *userPb.ListStorageIn
 		}
 		cids[i] = c
 	}
-	infos, err := i.ListStorageInfo(cids...)
+	infos, err := a.s.ListStorageInfo(iids, cids)
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "querying storage info: %v", err)
 	}
@@ -56,5 +55,5 @@ func (s *Service) ListStorageInfo(ctx context.Context, req *userPb.ListStorageIn
 	for i, info := range infos {
 		res[i] = su.ToRPCStorageInfo(info)
 	}
-	return &userPb.ListStorageInfoResponse{StorageInfo: res}, nil
+	return &adminPb.ListStorageInfoResponse{StorageInfo: res}, nil
 }
