+++ cli-docs/pow/pow_admin_storage-info.md
@@ -1,0 +1,28 @@
+## pow admin storage-info
+
+Provides admin storage info commands
+
+### Synopsis
+
+Provides admin storage info commands
+
+### Options
+
+```
+  -h, --help   help for storage-info
+```
+
+### Options inherited from parent commands
+
+```
+      --admin-token string     admin auth token
+      --serverAddress string   address of the powergate service api (default "127.0.0.1:5002")
+  -t, --token string           user auth token
+```
+
+### SEE ALSO
+
+* [pow admin](pow_admin.md)	 - Provides admin commands
+* [pow admin storage-info get](pow_admin_storage-info_get.md)	 - Returns the information about a stored cid.
+* [pow admin storage-info list](pow_admin_storage-info_list.md)	 - Returns a list of information about all stored cids, filtered by user ids and cids if provided.
+
