+++ cli-docs/pow/pow_admin_storage-info_get.md
@@ -1,0 +1,30 @@
+## pow admin storage-info get
+
+Returns the information about a stored cid.
+
+### Synopsis
+
+Returns the information about a stored cid.
+
+```
+pow admin storage-info get [user-id] [cid] [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for get
+```
+
+### Options inherited from parent commands
+
+```
+      --admin-token string     admin auth token
+      --serverAddress string   address of the powergate service api (default "127.0.0.1:5002")
+  -t, --token string           user auth token
+```
+
+### SEE ALSO
+
+* [pow admin storage-info](pow_admin_storage-info.md)	 - Provides admin storage info commands
+
